requirements.txt
jump-consistent-hash==3.1.2
